+++ .cargo/config.toml
@@ -1,0 +1,11 @@
+[target.wasm32-wasi]
+rustflags = [
+  "-Clink-arg=--export-table",
+  "-Clink-arg=--export=malloc",
+  "-Clink-arg=--export=free",
+  "-Clink-arg=-L/usr/lib/llvm-15/lib/clang/15.0.7/lib/wasi",
+  "-Clink-arg=-lclang_rt.builtins-wasm32"
+]
+
+[build]
+target = "wasm32-wasi"
