+++ .github/workflows/release.yml
@@ -13,9 +13,8 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v2
       - name: Build WASM module
-        run: |
-          ./src/wasm_navdata_interface/scripts/run_docker_cmd.sh ./scripts/build.sh
+        run: npm run build:wasm-workflow
       - name: Release
         uses: softprops/action-gh-release@v1
         with:
-          files: ./src/wasm_navdata_interface/out/navdata_interface.wasm
+          files: ./out/msfs_navdata_interface.wasm
