+++ .gitignore
@@ -5,8 +5,8 @@ _PackageInt
 tsconfig.tsbuildinfo
 .vs
 examples/aircraft/PackageSources/html_ui/Pages/VCockpit/Instruments/Navigraph/NavdataInterfaceSample
-examples/aircraft/PackageSources/SimObjects/Airplanes/Navigraph_Navdata_Interface_Aircraft/panel/navdata_interface.wasm
-src/wasm_navdata_interface/out
+examples/aircraft/PackageSources/SimObjects/Airplanes/Navigraph_Navdata_Interface_Aircraft/panel/msfs_navdata_interface.wasm
+out
 
 # Rust
 # will have compiled files and executables
