+++ CODEOWNERS
@@ -3,3 +3,4 @@
 # @Navigraph/app-dev will be requested for
 # review when someone opens a pull request.
 *       @pepperoni505
+*       @professoralex13
