+++ README.md
@@ -48,3 +48,12 @@ Before building, make sure you have properly created and set an `.env` file in `
 3. Run `.\build.bat` (must be on Windows)
    - This will take a while to download and build the first time, but subsequent runs will be quicker
 4. The compiled WASM module will be copied to `src/wasm_navdata_interface/out` **and** `examples/aircraft/PackageSources/SimObjects/Airplanes/Navigraph_Navdata_Interface_Aircraft/panel`
+
+# Building:
+
+- Install Docker
+- Run `.\build.bat` to build, and `.\run_cargo_cmd.bat` followed by a Cargo command to run a specified command (e.g `.\run_cargo_cmd.bat clippy`)
+
+# Warning
+
+The only file system functions that properly work are `fs::remove_file`, `fs::File::create`, `fs::create_dir`, `fs::remove_dir`, `fs::remove_file`. All the other functions regarding the filesystem do not work due to the MSFS implementation of WASI
