+++ package.json
@@ -2,12 +2,12 @@
   "name": "msfs-navdata-interface",
   "scripts": {
     "format": "prettier --write .",
-    "lint": "eslint --ext ts,tsx ."
+    "lint": "eslint --ext ts,tsx .",
+    "test": "ts-node -T src/test/test.ts",
+    
+    "build:wasm": ".\\scripts\\build.bat",
+    "build:wasm-workflow": "./scripts/build.sh"
   },
-  "workspaces": [
-    "src/*",
-    "examples/*"
-  ],
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
     "@microsoft/msfs-types": "^1.14.6",
@@ -18,6 +18,9 @@
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.1",
     "prettier": "^3.0.3",
-    "typescript": "^5.2.2"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.2.2",
+    "navigraph": "^1.2.35",
+    "random-bigint": "^0.0.1"
   }
 }
