+++ src/js/NavdataInterfaceTypes.ts
