--- src/navdata-sdk/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@navigraph/navdata-sdk",
-  "version": "0.0.1",
-  "license": "UNLICENSED",
-  "scripts": {
-    "test": "ts-node -T test.ts --experimental-wasi-unstable-preview1"
-  },
-  "dependencies": {
-    "navigraph": "^1.2.35",
-    "random-bigint": "^0.0.1",
-    "ts-node": "^10.9.2",
-    "uuid": "^9.0.1"
-  },
-  "devDependencies": {
-    "@types/uuid": "^9.0.7"
-  }
-}
