+++ src/wasm/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "msfs-navdata-interface"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+msfs = { git = "https://github.com/flybywiresim/msfs-rs.git", rev = "b438d3e" }
+serde = { version = "1.0.190", features = ["derive"] }
+serde_json = "1.0.108"
+rusqlite = { version = "0.30.0", features = ["bundled"] }
+zip = { version = "0.6.4", default-features = false, features = ["deflate"] }
+serde_rusqlite = "0.34.0"
