+++ src/wasm/src/download/mod.rs
@@ -1,0 +1,2 @@
+pub mod downloader;
+pub mod zip_handler;
