+++ src/wasm/src/query/mod.rs
