--- src/wasm_navdata_interface/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "navdata_interface"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.72.1"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-crate-type = ["cdylib"]
-
-[profile.release]
-lto = true
-strip = true
-
-[dependencies]
-msfs = { git = "https://github.com/flybywiresim/msfs-rs.git", rev = "b438d3e" }
-serde = { version = "1.0.190", features = ["derive"] }
-serde_json = "1.0.108"
-rusqlite = { version = "0.30.0", features = ["bundled"] }
-zip = { version = "0.6.4", default-features = false, features = ["deflate"] }
-serde_rusqlite = "0.34.0"
-
-[patch.crates-io]
-rusqlite = { git = "https://github.com/navigraph/rusqlite", rev = "7921774" }
