--- src/wasm_navdata_interface/rust-toolchain.toml
@@ -1,3 +1,0 @@
-[toolchain]
-profile = "default"
-channel = "1.72.1"
