--- src/wasm_navdata_interface/scripts/cargo_cmd.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-cargo "$@"
