--- src/wasm_navdata_interface/src/download/mod.rs
@@ -1,2 +1,0 @@
-pub mod downloader;
-pub mod zip_handler;
